Implement renaming of sansdata.py
